Convert SubmitProject action to Razor Pages method
We use Razor Pages for the majority of the app, and want to convert the remaining actions in the TransfersController. 

Refactor the SubmitProject action from Razor Views to the OnPost method of CheckYourAnswers
