+++ Frontend/Pages/Transfers/CheckYourAnswers.cshtml
@@ -118,7 +118,7 @@
                 <dd class="govuk-summary-list__actions"></dd>
             </div>
         </dl>
-        <form method="post" asp-action="SubmitProject" asp-controller="Transfers">
+        <form method="post">
             <button type="submit" data-module="govuk-button" data-prevent-double-click="true" class="govuk-button govuk-button-primary" data-test="create-project">Continue</button>
         </form>
     </div>
